Add dropDup and countFreq method in class HistogramGenerator body #4
* dropDup method creates an ArrayList with all the values of a single dimensional array that has been
created through a file form the createArray method. The values of this ArrayList will be used
as the values for x axis of the frequency histogram.

* countFreq creates an ArrayList with all the frequencies of the values of a single dimensional array.
The values of this ArrayList will be used as values for the y axis of the frequency histogram.

* The purpose of this change, in HistogramGenerator class, is to use the functionality of dropDup and
countFreq methods in order to generate a frequency histogram for a given file of grades.